to_irc_lower should account for channel names with '/' -- fixes #2
Channels are represented as filesystem directories
The latter cannot contain slashes ('/'), so they
are replaced with comma (',').
Channels are not allowed to have commas, so when
sending a command to a channel, the comma needs
to be converted back to a slash.
This has been added in the todo list.